refactored run-perf_tests.py to have less arguments and restructured benchmark folder (got rid of regular and large folders)
