+++ test/Performance/benchmark/pdo_sqlsrv/PDOConnectionBench.php
@@ -1,7 +1,9 @@
 <?php
 
 use PDOSqlsrvPerfTest\PDOSqlsrvUtil;
-
+/**
+ * @Iterations(1000)
+ */
 class PDOConnectionBench{
     /*
     * Opens a connection and closes it immediately
