+++ test/Performance/benchmark/pdo_sqlsrv/PDOCreateDbTableProcBench.php
@@ -1,31 +1,31 @@
 <?php
 
-use SqlsrvPerfTest\SqlsrvUtil;
+use PDOSqlsrvPerfTest\PDOSqlsrvUtil;
 /**
+ * @Iterations(1000)
  * @BeforeMethods({"connect"})
  * @AfterMethods({"disconnect"})
  */
-class SqlsrvCreateDbTableProcBench{
+class PDOCreateDbTableProcBench{
     private $conn;
 
     public function connect(){
-        $this->conn = SqlsrvUtil::connect();
+        $this->conn = PDOSqlsrvUtil::connect();
     }
-
     /*
     * Each iteration creates a database, a table and a stored procedure in that database and drops the database at the end.
-    * Note that, ODBC SQLExecDirect function are used to execute all the queries. 
+    * Note that, execDirect function are used to execute all the queries. 
     */
     public function benchCreateDbTableProc(){
         $randomNum = rand();
         $databaseName = "test_db_$randomNum";
         $tableName = "test_table_$randomNum";
         $procName = "test_proc_$randomNum";
-        SqlsrvUtil::createDbTableProc( $this->conn, $databaseName, $tableName, $procName );
-        SqlsrvUtil::dropDatabase( $this->conn, $databaseName );
+        PDOSqlsrvUtil::createDbTableProc( $this->conn, $databaseName, $tableName, $procName );
+        PDOSqlsrvUtil::dropDatabase( $this->conn, $databaseName );
     }
 
     public function disconnect(){
-        SqlsrvUtil::disconnect( $this->conn );
+        PDOSqlsrvUtil::disconnect( $this->conn );
     }
 }
