+++ test/Performance/benchmark/sqlsrv/SqlsrvConnectionBench.php
@@ -1,7 +1,9 @@
 <?php
 
 use SqlsrvPerfTest\SqlsrvUtil;
-
+/**
+ * @Iterations(1000)
+ */
 class SqlsrvConnectionBench{
     /*
     * Opens a connection and closes it immediately
