+++ test/Performance/benchmark/sqlsrv/SqlsrvCreateDbTableProcBench.php
@@ -1,30 +1,32 @@
 <?php
 
-use PDOSqlsrvPerfTest\PDOSqlsrvUtil;
+use SqlsrvPerfTest\SqlsrvUtil;
 /**
+ * @Iterations(1000)
  * @BeforeMethods({"connect"})
  * @AfterMethods({"disconnect"})
  */
-class PDOCreateDbTableProcBench{
+class SqlsrvCreateDbTableProcBench{
     private $conn;
 
     public function connect(){
-        $this->conn = PDOSqlsrvUtil::connect();
+        $this->conn = SqlsrvUtil::connect();
     }
+
     /*
     * Each iteration creates a database, a table and a stored procedure in that database and drops the database at the end.
-    * Note that, execDirect function are used to execute all the queries. 
+    * Note that, ODBC SQLExecDirect function are used to execute all the queries. 
     */
     public function benchCreateDbTableProc(){
         $randomNum = rand();
         $databaseName = "test_db_$randomNum";
         $tableName = "test_table_$randomNum";
         $procName = "test_proc_$randomNum";
-        PDOSqlsrvUtil::createDbTableProc( $this->conn, $databaseName, $tableName, $procName );
-        PDOSqlsrvUtil::dropDatabase( $this->conn, $databaseName );
+        SqlsrvUtil::createDbTableProc( $this->conn, $databaseName, $tableName, $procName );
+        SqlsrvUtil::dropDatabase( $this->conn, $databaseName );
     }
 
     public function disconnect(){
-        PDOSqlsrvUtil::disconnect( $this->conn );
+        SqlsrvUtil::disconnect( $this->conn );
     }
 }
