+++ test/Performance/lib/result_db.php
@@ -1,0 +1,6 @@
+<?php
+$server = 'server';
+$database = 'testdb';
+$uid = 'usr';
+$pwd = 'pwd';
+?>
